Exercise 4.21

+++ part4/blog/controllers/blogs.js
@@ -34,7 +34,14 @@ blogRouter.post('/', async (request, response) => {
 });
 
 blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id);
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!decodedToken) {
+        return response.status(401).json({ error: 'token invalid' });
+    }
+    const blogToDelete = await Blog.findById(request.params.id);
+    if (blogToDelete && blogToDelete.user.toString() === decodedToken.id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id);
+    }
     response.status(204).end();
 });
